First commit of the redundantptr analyser
This software checks golang code to find if a struct field is populated
by a pointer to a variable. This maybe somethig that can be replaced by
`ptr.To()` function.

This is not 100% accurate. sometimes it doesn't make sense to replace.
e.g. in this case, we want the `a` variable because there is a logic
arounf it.

```golang
var a = "something"
if something happened {
	a = "something else"
}

x := SomeStruct{
	somePointerField: &a,
}
```